122 - separate data source config from main config file and test field_spec_mapping_to_jq_field_spec_mapping
