+++ tel2puml/otel_to_pv/data_sources/data_sources_config.py
@@ -1,0 +1,10 @@
+"""Module for DataSources typed dict."""
+from typing import TypedDict, NotRequired
+
+from .json_data_source.json_config import JSONDataSourceConfig
+
+
+class DataSources(TypedDict):
+    """Typed dict for DataSources."""
+
+    json: NotRequired[JSONDataSourceConfig]
