+++ tel2puml/otel_to_pv/data_sources/json_data_source/json_jq_converter.py
@@ -5,7 +5,7 @@ from typing import Any, Generator
 
 import jq  # type: ignore[import-not-found]
 
-from tel2puml.otel_to_pv.config import (
+from .json_config import (
     JQFieldSpec,
     field_spec_mapping_to_jq_field_spec_mapping,
     FieldSpec,
